Separate BuildGate waiter from File system change waiter
